combat manager managing
This commit adds command classes (not in use yet).
Also, the on_cell_change edits to combatants that were previously made in directly in the main_window are now handled by the combat manager.

+++ models/combatant.py
@@ -9,7 +9,7 @@ class Ability:
 @dataclass
 class Combatant:
     name: str
-    initiative: int
+    initiative: float
     ac: int
     hp_total: int
     #Setting up an id to use for identification.
